+++ packages/frontend/package.json
@@ -19,8 +19,7 @@
     "@unirep/core": "2.0.0-alpha-2",
     "ethers": "^5.7.2",
     "file-loader": "^6.2.0",
-    "mobx-react-lite": "^3.4.0",
-    "react-icons": "^4.7.1"
+    "mobx-react-lite": "^3.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.19.6",
