+++ packages/frontend/public/index.html
@@ -10,6 +10,7 @@
             content=""
         />
         <link rel="apple-touch-icon" href="logo192.png" />
+        <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Azeret+Mono">
         <title>unirep app</title>
         <script>
             window.__DEV_CONFIG__ = '__DEV_CONFIG__'
