+++ packages/frontend/src/components/Logo.jsx
@@ -1,0 +1,31 @@
+import React from "react"
+
+export default () => {
+    return (
+        <svg width="185" height="48" viewBox="0 0 185 48" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <g clip-path="url(#clip0_512_1287)">
+                <path fill-rule="evenodd" clip-rule="evenodd" d="M1.2 0C0.537258 0 0 0.537257 0 1.2V29.9294C0 39.9095 8.09048 48 18.0706 48C18.6943 48 19.2 47.4943 19.2 46.8706V1.2C19.2 0.537258 18.6627 0 18 0H1.2ZM9.6 2.4C9.6 6.37645 6.37645 9.6 2.4 9.6C6.37645 9.6 9.6 12.8236 9.6 16.8C9.6 12.8236 12.8235 9.6 16.8 9.6C12.8235 9.6 9.6 6.37645 9.6 2.4Z" fill="#A0D8DB"/>
+                <path fill-rule="evenodd" clip-rule="evenodd" d="M22.8 48C22.1372 48 21.6 47.4627 21.6 46.8L21.6 18.0706C21.6 8.09048 29.6905 8.72489e-07 39.6706 0C40.2943 -5.45306e-08 40.8 0.505655 40.8 1.12941L40.8 46.8C40.8 47.4627 40.2627 48 39.6 48H22.8ZM31.2 45.6C31.2 41.6236 27.9764 38.4 24 38.4C27.9764 38.4 31.2 35.1765 31.2 31.2C31.2 35.1765 34.4235 38.4 38.4 38.4C34.4235 38.4 31.2 41.6236 31.2 45.6Z" fill="#A0D8DB"/>
+            </g>
+            <path d="M57.984 26.72C56.0427 26.72 54.3893 26.3787 53.024 25.696C51.6587 24.992 50.6133 24.0107 49.888 22.752C49.1627 21.4933 48.8 20.0213 48.8 18.336V4H54.08V18.336C54.08 19.168 54.24 19.8827 54.56 20.48C54.88 21.0773 55.328 21.5467 55.904 21.888C56.5013 22.208 57.1947 22.368 57.984 22.368C58.7947 22.368 59.488 22.208 60.064 21.888C60.64 21.5467 61.088 21.0773 61.408 20.48C61.728 19.8827 61.888 19.168 61.888 18.336V4H67.168V18.336C67.168 20.0213 66.8053 21.4933 66.08 22.752C65.3547 24.0107 64.3093 24.992 62.944 25.696C61.5787 26.3787 59.9253 26.72 57.984 26.72Z" fill="#151616"/>
+            <path d="M72.5885 26.336V4H79.3085L85.5485 19.808L85.1965 11.68V4H90.1565V26.336H83.4365L77.1325 10.816L77.5485 18.656V26.336H72.5885Z" fill="#151616"/>
+            <path d="M102.073 26.336V4H107.353V26.336H102.073ZM96.953 4H112.473V8.32H96.953V4ZM96.953 22.016H112.473V26.336H96.953V22.016Z" fill="#151616"/>
+            <path d="M123.493 17.984V13.664H128.517C129.157 13.664 129.712 13.5467 130.181 13.312C130.672 13.0773 131.045 12.7467 131.301 12.32C131.579 11.8933 131.717 11.4133 131.717 10.88C131.717 10.368 131.589 9.92 131.333 9.536C131.099 9.152 130.768 8.85333 130.341 8.64C129.915 8.42667 129.424 8.32 128.869 8.32H123.493V4H128.869C130.491 4 131.92 4.27733 133.157 4.832C134.416 5.36533 135.397 6.144 136.101 7.168C136.805 8.192 137.157 9.42933 137.157 10.88C137.157 11.7547 136.944 12.6187 136.517 13.472C136.112 14.304 135.515 15.0613 134.725 15.744C133.957 16.4267 133.019 16.9707 131.909 17.376C130.821 17.7813 129.584 17.984 128.197 17.984H123.493ZM119.685 26.336V4H124.965V26.336H119.685ZM131.845 26.336C131.845 25.3973 131.675 24.448 131.333 23.488C130.992 22.5067 130.491 21.5787 129.829 20.704C129.189 19.808 128.389 18.9973 127.429 18.272C126.469 17.5467 125.36 16.96 124.101 16.512L129.957 15.328C131.195 15.9467 132.272 16.6827 133.189 17.536C134.107 18.3893 134.864 19.3067 135.461 20.288C136.08 21.2693 136.539 22.2827 136.837 23.328C137.136 24.352 137.285 25.3547 137.285 26.336H131.845Z" fill="#151616"/>
+            <path d="M143.41 26.336V4H159.954V8.32H148.69V12.64H155.954V16.96H148.69V22.016H160.082V26.336H143.41Z" fill="#151616"/>
+            <path d="M170.142 19.328V15.008H175.23C176.254 15.008 177.076 14.6987 177.694 14.08C178.313 13.44 178.622 12.608 178.622 11.584C178.622 10.6027 178.313 9.81333 177.694 9.216C177.076 8.61867 176.254 8.32 175.23 8.32H170.142V4H175.262C177.033 4 178.58 4.30933 179.902 4.928C181.225 5.54667 182.249 6.432 182.974 7.584C183.7 8.71467 184.062 10.048 184.062 11.584C184.062 13.1413 183.7 14.5067 182.974 15.68C182.249 16.832 181.225 17.728 179.902 18.368C178.601 19.008 177.054 19.328 175.262 19.328H170.142ZM166.366 26.336V4H171.646V26.336H166.366Z" fill="#151616"/>
+            <path d="M50.584 38.996V37.496H54.384C55.3573 37.496 56.144 37.236 56.744 36.716C57.344 36.196 57.644 35.5093 57.644 34.656C57.644 33.816 57.364 33.1427 56.804 32.636C56.2573 32.1293 55.5307 31.876 54.624 31.876H50.584V30.376H54.644C55.6173 30.376 56.464 30.556 57.184 30.916C57.904 31.2627 58.464 31.756 58.864 32.396C59.264 33.0227 59.464 33.776 59.464 34.656C59.464 35.5227 59.2507 36.2827 58.824 36.936C58.4107 37.5893 57.824 38.096 57.064 38.456C56.304 38.816 55.4107 38.996 54.384 38.996H50.584ZM49.344 44.336V30.376H51.064V44.336H49.344Z" fill="#151616"/>
+            <path d="M68.4718 38.296V36.796H72.1318C72.7851 36.796 73.3585 36.696 73.8518 36.496C74.3585 36.2827 74.7518 35.9827 75.0318 35.596C75.3251 35.196 75.4718 34.736 75.4718 34.216C75.4718 33.736 75.3451 33.3227 75.0918 32.976C74.8385 32.6293 74.4985 32.3627 74.0718 32.176C73.6451 31.976 73.1718 31.876 72.6518 31.876H68.4718V30.376H72.6518C73.5451 30.376 74.3385 30.5293 75.0318 30.836C75.7385 31.1427 76.2918 31.5827 76.6918 32.156C77.0918 32.7293 77.2918 33.416 77.2918 34.216C77.2918 34.7493 77.1718 35.2627 76.9318 35.756C76.7051 36.236 76.3651 36.6693 75.9118 37.056C75.4585 37.4293 74.9118 37.7293 74.2718 37.956C73.6318 38.1827 72.8985 38.296 72.0718 38.296H68.4718ZM67.2318 44.336V30.376H68.9518V44.336H67.2318ZM75.3918 44.336C75.2851 43.696 75.0918 43.0693 74.8118 42.456C74.5451 41.8427 74.1851 41.256 73.7318 40.696C73.2918 40.136 72.7385 39.6093 72.0718 39.116C71.4051 38.6227 70.6318 38.1693 69.7518 37.756L71.6718 37.316C72.6051 37.7693 73.3985 38.2693 74.0518 38.816C74.7051 39.3493 75.2451 39.916 75.6718 40.516C76.1118 41.1027 76.4451 41.716 76.6718 42.356C76.8985 42.996 77.0518 43.656 77.1318 44.336H75.3918Z" fill="#151616"/>
+            <path d="M89.7196 44.576C88.5463 44.576 87.5263 44.2827 86.6596 43.696C85.793 43.096 85.1196 42.256 84.6396 41.176C84.173 40.0827 83.9396 38.8093 83.9396 37.356C83.9396 35.8893 84.173 34.616 84.6396 33.536C85.1196 32.456 85.793 31.6227 86.6596 31.036C87.5263 30.436 88.5463 30.136 89.7196 30.136C90.893 30.136 91.913 30.436 92.7796 31.036C93.6463 31.6227 94.313 32.456 94.7796 33.536C95.2596 34.616 95.4996 35.8893 95.4996 37.356C95.4996 38.8093 95.2596 40.0827 94.7796 41.176C94.313 42.256 93.6463 43.096 92.7796 43.696C91.913 44.2827 90.893 44.576 89.7196 44.576ZM89.7196 43.076C90.5196 43.076 91.213 42.8427 91.7996 42.376C92.3996 41.9093 92.8596 41.2493 93.1796 40.396C93.513 39.5293 93.6796 38.516 93.6796 37.356C93.6796 36.196 93.513 35.1893 93.1796 34.336C92.8596 33.4827 92.3996 32.8227 91.7996 32.356C91.213 31.876 90.5196 31.636 89.7196 31.636C88.9196 31.636 88.2196 31.876 87.6196 32.356C87.033 32.8227 86.573 33.4827 86.2396 34.336C85.9063 35.1893 85.7396 36.196 85.7396 37.356C85.7396 38.516 85.9063 39.5293 86.2396 40.396C86.573 41.2493 87.033 41.9093 87.6196 42.376C88.2196 42.8427 88.9196 43.076 89.7196 43.076Z" fill="#151616"/>
+            <path d="M106.627 44.336V31.876H101.987V30.376H113.007V31.876H108.367V44.336H106.627Z" fill="#151616"/>
+            <path d="M125.335 44.576C124.162 44.576 123.142 44.2827 122.275 43.696C121.409 43.096 120.735 42.256 120.255 41.176C119.789 40.0827 119.555 38.8093 119.555 37.356C119.555 35.8893 119.789 34.616 120.255 33.536C120.735 32.456 121.409 31.6227 122.275 31.036C123.142 30.436 124.162 30.136 125.335 30.136C126.509 30.136 127.529 30.436 128.395 31.036C129.262 31.6227 129.929 32.456 130.395 33.536C130.875 34.616 131.115 35.8893 131.115 37.356C131.115 38.8093 130.875 40.0827 130.395 41.176C129.929 42.256 129.262 43.096 128.395 43.696C127.529 44.2827 126.509 44.576 125.335 44.576ZM125.335 43.076C126.135 43.076 126.829 42.8427 127.415 42.376C128.015 41.9093 128.475 41.2493 128.795 40.396C129.129 39.5293 129.295 38.516 129.295 37.356C129.295 36.196 129.129 35.1893 128.795 34.336C128.475 33.4827 128.015 32.8227 127.415 32.356C126.829 31.876 126.135 31.636 125.335 31.636C124.535 31.636 123.835 31.876 123.235 32.356C122.649 32.8227 122.189 33.4827 121.855 34.336C121.522 35.1893 121.355 36.196 121.355 37.356C121.355 38.516 121.522 39.5293 121.855 40.396C122.189 41.2493 122.649 41.9093 123.235 42.376C123.835 42.8427 124.535 43.076 125.335 43.076Z" fill="#151616"/>
+            <path d="M143.403 44.576C142.176 44.576 141.11 44.2627 140.203 43.636C139.31 43.0093 138.623 42.156 138.143 41.076C137.663 39.9827 137.423 38.736 137.423 37.336C137.423 35.8693 137.676 34.596 138.183 33.516C138.69 32.436 139.396 31.6027 140.303 31.016C141.21 30.4293 142.256 30.136 143.443 30.136C144.416 30.136 145.29 30.3293 146.063 30.716C146.836 31.0893 147.47 31.6293 147.963 32.336C148.456 33.0427 148.75 33.876 148.843 34.836H146.923C146.816 33.876 146.45 33.1027 145.823 32.516C145.196 31.9293 144.396 31.636 143.423 31.636C142.596 31.636 141.87 31.8693 141.243 32.336C140.616 32.8027 140.123 33.4693 139.763 34.336C139.416 35.1893 139.243 36.196 139.243 37.356C139.243 38.5293 139.423 39.5493 139.783 40.416C140.143 41.2693 140.636 41.9293 141.263 42.396C141.89 42.8493 142.603 43.076 143.403 43.076C144.083 43.076 144.676 42.936 145.183 42.656C145.703 42.376 146.116 41.9827 146.423 41.476C146.743 40.9693 146.936 40.376 147.003 39.696H148.903C148.81 40.6693 148.523 41.5293 148.043 42.276C147.576 43.0093 146.95 43.576 146.163 43.976C145.376 44.376 144.456 44.576 143.403 44.576Z" fill="#151616"/>
+            <path d="M160.951 44.576C159.778 44.576 158.758 44.2827 157.891 43.696C157.024 43.096 156.351 42.256 155.871 41.176C155.404 40.0827 155.171 38.8093 155.171 37.356C155.171 35.8893 155.404 34.616 155.871 33.536C156.351 32.456 157.024 31.6227 157.891 31.036C158.758 30.436 159.778 30.136 160.951 30.136C162.124 30.136 163.144 30.436 164.011 31.036C164.878 31.6227 165.544 32.456 166.011 33.536C166.491 34.616 166.731 35.8893 166.731 37.356C166.731 38.8093 166.491 40.0827 166.011 41.176C165.544 42.256 164.878 43.096 164.011 43.696C163.144 44.2827 162.124 44.576 160.951 44.576ZM160.951 43.076C161.751 43.076 162.444 42.8427 163.031 42.376C163.631 41.9093 164.091 41.2493 164.411 40.396C164.744 39.5293 164.911 38.516 164.911 37.356C164.911 36.196 164.744 35.1893 164.411 34.336C164.091 33.4827 163.631 32.8227 163.031 32.356C162.444 31.876 161.751 31.636 160.951 31.636C160.151 31.636 159.451 31.876 158.851 32.356C158.264 32.8227 157.804 33.4827 157.471 34.336C157.138 35.1893 156.971 36.196 156.971 37.356C156.971 38.516 157.138 39.5293 157.471 40.396C157.804 41.2493 158.264 41.9093 158.851 42.376C159.451 42.8427 160.151 43.076 160.951 43.076Z" fill="#151616"/>
+            <path d="M174.619 44.336V30.376H176.339V42.836H183.839V44.336H174.619Z" fill="#151616"/>
+            <defs>
+                <clipPath id="clip0_512_1287">
+                <rect width="40.8" height="48" fill="white"/>
+                </clipPath>
+            </defs>
+        </svg>
+    )
+}
