+++ packages/frontend/src/components/infoModal.css
@@ -1,0 +1,14 @@
+.modal {
+    background-color: #456601;
+    color: white;
+    font-size: 16px;
+    font-weight: 500;
+    border: 2px;
+    border-color: white;
+    border-radius: 10px;
+    width: 200px;
+    height: auto;
+    padding: 20px;
+    z-index: 10;
+    position: absolute;
+}
