+++ packages/frontend/src/index.css
@@ -1,3 +1,4 @@
 body {
-  font-family: sans-serif, helvetica;
+  /* font-family: sans-serif, helvetica; */
+  font-family: 'Azeret Mono';
 }
