+++ packages/frontend/src/index.jsx
@@ -1,7 +1,23 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import Home from './Home'
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Header from './pages/Header';
+import Start from './pages/Start'
+import Dashboard from './pages/Dashboard'
 import './index.css'
 
+export default function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Header />}>
+          <Route index element={<Start />} />
+          <Route path="dashboard" element={<Dashboard />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
 const root = createRoot(document.getElementById('root'))
-root.render(<Home />)
+root.render(<App />)
