+++ packages/frontend/src/pages/header.css
@@ -1,0 +1,20 @@
+a:link {
+    margin-left: 80px;
+    font-weight: 600;
+    color: black;
+    text-decoration: none;
+}
+
+a:visited {
+    color: black;
+}
+
+.header {
+    display: flex;
+    justify-content: space-between;
+    margin: 50px;
+}
+
+.links {
+    display: flex;
+}
